+++ src/ui/mainwindow/chartwidget.cpp
@@ -289,6 +289,7 @@ Charts ChartWidget::addModel(QSharedPointer<AbstractModel> model)
         m_signal_y = model->YLabel();
         m_error_x = model->XLabel();
         m_error_y = "Error " + model->YLabel();
+        Repaint();
     });
 
     m_signal_x = model->XLabel();