+++ src/ui/widgets/systemparameterwidget.cpp
@@ -36,16 +36,18 @@ SystemParameterWidget::SystemParameterWidget(const SystemParameter& parameter, b
 
 {
     m_textfield = new QLineEdit;
-    m_textfield->setReadOnly(m_readonly);
+    // m_textfield->setReadOnly(m_readonly);
     m_boolbox = new QCheckBox;
-    m_boolbox->setDisabled(m_readonly);
+    // m_boolbox->setDisabled(m_readonly);
     m_list = new QComboBox;
-    m_list->setDisabled(m_readonly);
+    m_list->setItemDelegate(new HTMLDelegate());
+    // m_list->setDisabled(m_readonly);
 
     QLabel* label = new QLabel(parameter.Description());
     label->setFixedWidth(250);
     connect(m_textfield, SIGNAL(textChanged(QString)), this, SLOT(PrepareChanged()));
     connect(m_boolbox, SIGNAL(stateChanged(int)), this, SLOT(PrepareChanged()));
+    connect(m_list, SIGNAL(currentIndexChanged(int)), this, SLOT(PrepareChanged()));
 
     setTitle(parameter.Name());