+++ src/ui/widgets/systemparameterwidget.h
@@ -21,12 +21,51 @@
 
 #include "src/core/AbstractModel.h"
 
+#include <QtGui/QPainter>
+#include <QtGui/QStyledItemDelegate>
+#include <QtGui/QTextDocument>
+
 #include <QtWidgets/QCheckBox>
 #include <QtWidgets/QComboBox>
 #include <QtWidgets/QDoubleSpinBox>
 #include <QtWidgets/QGroupBox>
 #include <QtWidgets/QLayout>
 #include <QtWidgets/QLineEdit>
+
+class HTMLDelegate : public QStyledItemDelegate {
+protected:
+    void paint(QPainter* painter, const QStyleOptionViewItem& option, const QModelIndex& index) const
+    {
+        QStyleOptionViewItemV4 options = option;
+        initStyleOption(&options, index);
+
+        painter->save();
+
+        QTextDocument doc;
+        doc.setHtml(options.text);
+
+        options.text = "";
+        options.widget->style()->drawControl(QStyle::CE_ItemViewItem, &options, painter);
+        options.widget->style()->drawControl(QStyle::CE_ComboBoxLabel, &options, painter);
+
+        painter->translate(options.rect.left(), options.rect.top());
+        QRect clip(0, 0, options.rect.width(), options.rect.height());
+        doc.drawContents(painter, clip);
+
+        painter->restore();
+    }
+
+    QSize sizeHint(const QStyleOptionViewItem& option, const QModelIndex& index) const
+    {
+        QStyleOptionViewItemV4 options = option;
+        initStyleOption(&options, index);
+
+        QTextDocument doc;
+        doc.setHtml(options.text);
+        doc.setTextWidth(options.rect.width());
+        return QSize(doc.idealWidth(), doc.size().height());
+    }
+};
 
 class SystemParameterWidget : public QGroupBox {
     Q_OBJECT